redid locking + signaling
The same lock must be used for signaling the change in the
global_restart* flags and the `onDataCallback` fields in the video
and audio streamers.  We use `mutex_main` for this purpose.

This is on an infrequently executing path when the streaming threads
are going into and out of wait.

To address the issue of making sure that a streamer thread is not
executing `onDataCallback` while another thread deallocates the IMPDeviceSource
object, the original stream lock was repurposed to just protect onDataCallback
from deallocation.

In addition, an atomic `hasDataCallback` flag was introduced to allow the
streamer threads to check (without data races) whether they might skip
grabbing one frame.  (This is not 100% accurate: they might see `hasDataCallback`
to be false when it's already true, but this just means we'll lose one frame.)

This way, on the common path, the streamer threads will only 2 atomic reads
of `hasDataCallback` and one (likely uncontended) lock acquisition of their
own stream lock.